deed-reader-web/backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# AI APIs
openai==1.6.1
tiktoken==0.5.2
anthropic==0.21.3

# Document processing
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.2.0
python-magic==0.4.27

# OCR capabilities
pytesseract==0.3.10
PyMuPDF==1.23.8
opencv-python==4.8.1.78

# Data processing
numpy==1.26.2
pandas==2.1.4

# Request handling
requests==2.31.0
urllib3==2.1.0
httpx==0.27.2

# Environment management
python-dotenv==1.0.0

# Validation
marshmallow==3.20.2

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# Development
python-json-logger==2.0.7

# Linting and formatting
flake8==7.0.0
black==24.3.0
isort==5.13.2
pylint==3.0.3
mypy==1.8.0
pre-commit==3.6.0